refactor route actions
