+++ src/client/src/containers/PageNewProject/QuickStart/index.tsx
@@ -20,7 +20,7 @@ import { sendTelemetry } from "../../../utils/extensionService/extensionService"
 import { setSelectedFrontendFrameworkAction, setSelectedBackendFrameworkAction } from "../../../store/userSelection/frameworks/action";
 import { setPagesAction } from "../../../store/userSelection/pages/action";
 import { AppContext } from "../../../AppContext";
-import { setVisitedWizardPageAction, setPageWizardPageAction } from "../../../store/config/pages/action";
+import { setVisitedWizardPageAction, setPageWizardPageAction } from "../../../store/navigation/routes/action";
 
 interface IStateProps {
   isPreview: boolean;
