+++ src/client/src/store/config/combineReducers.ts
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import detailsPage from "./pages/detailsPageReducer";
+import detailsPage from "../navigation/routes/detailsPageReducer";
 import previewStatus from "./config/previewReducer";
 import validations from "./validations/setValidationsReducer";
 import isValidatingName from "./validations/validatingNameReducer";
