+++ src/client/src/store/config/configType.ts
@@ -1,6 +1,6 @@
 import { IVersionData } from "./versions/model";
 
-import { ISetVisitedPageAction, IResetVisitedPageAction, ISetPageAction, ISetDetails } from "./pages/model";
+import { ISetVisitedPageAction, IResetVisitedPageAction, ISetPageAction, ISetDetails } from "../navigation/routes/model";
 import { IResetWizardAction } from "./config/model";
 
 type WizardInfoType =
